Испрвали неймениг классов
