+++ src/PhotoCatalog.Domain/Primitives/ResultVoid.cs
@@ -1,0 +1,54 @@
+﻿using System;
+
+namespace PhotoCatalog.Domain.Primitives;
+
+/// <summary>
+/// Представляет результат выполнения операции, не возвращающей значение (аналог void).
+/// Инкапсулирует логику успеха или провала.
+/// </summary>
+/// <remarks>
+/// ВНИМАНИЕ: Не создавайте объект через конструктор по умолчанию.
+/// Для инициализации используйте исключительно статические методы: <see cref="Success"/> или <see cref="Failure"/>.
+/// </remarks>
+public readonly record struct ResultVoid
+{
+    /// <summary>
+    /// Указывает, завершилась ли операция успешно.
+    /// </summary>
+    public bool IsSuccess { get; }
+
+    /// <summary>
+    /// Указывает, завершилась ли операция с ошибкой.
+    /// </summary>
+    public bool IsFailure => !IsSuccess;
+
+    /// <summary>
+    /// Объект детализированной ошибки. Если операция успешна, содержит <see cref="Primitives.Error.None"/>.
+    /// </summary>
+    public Error Error { get; }
+
+
+    /// <summary>
+    /// Приватный конструктор для инициализации внутреннего состояния.
+    /// </summary>
+    /// <param name="isSuccess">Флаг успешности операции.</param>
+    /// <param name="error">Объект ошибки.</param>
+    private ResultVoid(bool isSuccess, Error error)
+    {
+        IsSuccess = isSuccess;
+        Error = error;
+    }
+
+    /// <summary>
+    /// Создает успешный результат без ошибок.
+    /// </summary>
+    /// <returns>Успешный <see cref="ResultVoid"/>, где <see cref="IsSuccess"/> равно true, а <see cref="Error"/> равно <see cref="Primitives.Error.None"/>.</returns>
+    public static ResultVoid Success() => new(true, Error.None);
+
+    /// <summary>
+    /// Создает провальный результат с указанной ошибкой.
+    /// </summary>
+    /// <param name="error">Бизнес-ошибка, объясняющая причину провала.</param>
+    /// <returns>Провальный <see cref="ResultVoid"/>, где <see cref="IsFailure"/> равно true, а <see cref="Error"/> содержит переданную ошибку.</returns>
+    public static ResultVoid Failure(Error error) => new(false, error);
+}
