+++ test/PhotoCatalogUnit.Test/ResultVoidTest.cs
@@ -1,0 +1,42 @@
+﻿using PhotoCatalog.Domain.Primitives;
+
+using Xunit;
+
+namespace PhotoCatalogUnit.Test;
+
+/// <summary>
+/// Тесты для проверки поведения базового объекта <see cref="ResultVoid"/>.
+/// </summary>
+public class ResultVoidTests
+{
+    /// <summary>
+    /// Проверяет, что вызов <see cref="ResultVoid.Success"/> корректно инициализирует успешное состояние: 
+    /// флаг <see cref="ResultVoid.IsSuccess"/> устанавливается в true, а свойство ошибки содержит <see cref="Error.None"/>.
+    /// </summary>
+    [Fact]
+    public void Success_ShouldReturnIsSuccessTrue_And_ErrorNone()
+    {
+        var result = ResultVoid.Success();
+
+        Assert.True(result.IsSuccess);
+        Assert.False(result.IsFailure);
+        Assert.Equal(Error.None, result.Error);
+    }
+
+    /// <summary>
+    /// Проверяет, что вызов <see cref="ResultVoid.Failure"/> корректно инициализирует провальное состояние: 
+    /// флаг <see cref="ResultVoid.IsFailure"/> устанавливается в true, а свойство <see cref="ResultVoid.Error"/> 
+    /// строго соответствует переданному объекту бизнес-ошибки.
+    /// </summary>
+    [Fact]
+    public void Failure_WithValidError_ShouldReturnIsFailureTrue_And_MatchError()
+    {
+        var expectedError = new Error("Test.Failure", "Тестовое сообщение об ошибке");
+
+        var result = ResultVoid.Failure(expectedError);
+
+        Assert.False(result.IsSuccess);
+        Assert.True(result.IsFailure);
+        Assert.Equal(expectedError, result.Error);
+    }
+}
